+++ cedar-testing/tests/cedar-policy/decimal.rs
@@ -15,10 +15,8 @@
  */
 
 //! Integration tests targeting the decimal extension
-//!
-//! These files exist separately in the `CedarIntegrationTests` package
 
-use super::perform_integration_test_from_json;
+use cedar_testing::integration_testing::perform_integration_test_from_json;
 use std::path::Path;
 
 /// Path of the folder containing the JSON tests