+++ cedar-testing/tests/cedar-policy/multi.rs
@@ -14,31 +14,37 @@
  * limitations under the License.
  */
 
-//! Integration tests targeting the ipaddr extension
-//!
+//! Integration tests that involve interactions between multiple policies
 
-use cedar_policy::integration_testing::perform_integration_test_from_json;
+use cedar_testing::integration_testing::perform_integration_test_from_json;
 use std::path::Path;
 
 /// Path of the folder containing the JSON tests
 fn folder() -> &'static Path {
-    Path::new("tests/ip")
+    Path::new("tests/multi")
 }
 
 #[test]
-#[cfg(feature = "ipaddr")]
-fn ip_1() {
+fn multi_1() {
     perform_integration_test_from_json(folder().join("1.json"));
 }
 
 #[test]
-#[cfg(feature = "ipaddr")]
-fn ip_2() {
+fn multi_2() {
     perform_integration_test_from_json(folder().join("2.json"));
 }
 
 #[test]
-#[cfg(feature = "ipaddr")]
-fn ip_3() {
+fn multi_3() {
     perform_integration_test_from_json(folder().join("3.json"));
+}
+
+#[test]
+fn multi_4() {
+    perform_integration_test_from_json(folder().join("4.json"));
+}
+
+#[test]
+fn multi_5() {
+    perform_integration_test_from_json(folder().join("5.json"));
 }