+++ cedar-wasm/src/policies_and_templates.rs
@@ -1,6 +1,6 @@
 use std::str::FromStr;
 
-use cedar_policy::{Policy, PolicySet};
+use cedar_policy::{Policy, PolicySet, Template};
 use cedar_policy_core::parser::parse_policy_or_template_to_est;
 use serde::{Deserialize, Serialize};
 
@@ -8,10 +8,13 @@ use tsify::Tsify;
 use wasm_bindgen::prelude::*;
 
 #[derive(Tsify, Debug, Serialize, Deserialize)]
+#[serde(tag = "success")]
 #[serde(rename_all = "camelCase")]
 #[tsify(into_wasm_abi, from_wasm_abi)]
 pub enum JsonToPolicyResult {
+    #[serde(rename = "true")]
     Success { policy_text: String },
+    #[serde(rename = "false")]
     Error { errors: Vec<String> },
 }
 
@@ -37,15 +40,15 @@ pub fn policy_text_from_json(json_str: &str) -> JsonToPolicyResult {
 }
 
 #[derive(Tsify, Debug, Serialize, Deserialize, Clone)]
-#[serde(untagged)]
+#[serde(tag = "success")]
 #[tsify(into_wasm_abi, from_wasm_abi)]
 pub enum PolicyToJsonResult {
+    #[serde(rename = "true")]
     Success {
         policy: cedar_policy_core::est::Policy,
     },
-    Error {
-        errors: Vec<String>,
-    },
+    #[serde(rename = "false")]
+    Error { errors: Vec<String> },
 }
 
 #[wasm_bindgen(js_name = "policyTextToJson")]
@@ -59,11 +62,14 @@ pub fn policy_text_to_json(cedar_str: &str) -> PolicyToJsonResult {
 }
 
 #[derive(Tsify, Debug, Serialize, Deserialize)]
+#[serde(tag = "success")]
 #[tsify(into_wasm_abi, from_wasm_abi)]
 /// struct that defines the result for the syntax validation function
 pub enum CheckParsePolicySetResult {
+    #[serde(rename = "true")]
     /// represents successful syntax validation
     Success { policies: i32, templates: i32 },
+    #[serde(rename = "false")]
     /// represents a syntax error and encloses a vector of the errors
     SyntaxError { errors: Vec<String> },
 }
@@ -92,18 +98,40 @@ pub fn check_parse_policy_set(input_policies_str: &str) -> CheckParsePolicySetRe
     }
 }
 
-#[derive(Debug, Serialize, Deserialize)]
-#[wasm_bindgen]
-pub struct Template {
-    text: String,
-    slots: Vec<String>,
-    parse_errors: Option<Vec<String>>,
+#[derive(Tsify, Debug, Serialize, Deserialize)]
+#[serde(tag = "success")]
+#[tsify(into_wasm_abi, from_wasm_abi)]
+pub enum CheckParseTemplateResult {
+    #[serde(rename = "true")]
+    /// represents successful template validation
+    Success { slots: Vec<String> },
+    #[serde(rename = "false")]
+    /// represents errors in the template validation and encloses a vector of the errors
+    Error { errors: Vec<String> },
+}
+
+#[wasm_bindgen(js_name = "checkParseTemplate")]
+pub fn check_parse_template(template_str: &str) -> CheckParseTemplateResult {
+    match Template::from_str(template_str) {
+        Err(parse_errs) => CheckParseTemplateResult::Error {
+            errors: parse_errs.errors_as_strings(),
+        },
+        Ok(template) => match template.slots().count() {
+            1 | 2 => CheckParseTemplateResult::Success {
+                slots: template.slots().map(|slot| slot.to_string()).collect(),
+            },
+            _ => CheckParseTemplateResult::Error {
+                errors: vec!["Expected template to have one or two slots".to_string()],
+            },
+        },
+    }
 }
 
 #[cfg(test)]
 mod test {
 
     use super::*;
+    use cool_asserts::assert_matches;
 
     #[test]
     fn test_conversion_from_cedar() {
@@ -191,6 +219,35 @@ mod test {
         assert!(matches!(
             result,
             CheckParsePolicySetResult::SyntaxError { errors: _ }
+        ));
+    }
+
+    #[test]
+    fn can_parse_template() {
+        let template_str = r#"permit (principal == ?principal, action, resource == ?resource);"#;
+        let result = check_parse_template(template_str);
+        assert_matches!(result, CheckParseTemplateResult::Success { slots } => {
+            assert_eq!(slots.len(), 2);
+        });
+    }
+
+    #[test]
+    fn parse_template_fails_for_missing_slots() {
+        let template_str = r#"permit (principal, action, resource);"#;
+        let result = check_parse_template(template_str);
+        assert!(matches!(
+            result,
+            CheckParseTemplateResult::Error { errors: _ }
+        ));
+    }
+
+    #[test]
+    fn parse_template_fails_for_bad_slot() {
+        let template_str = r#"permit (principal, action, resource == ?principal);"#;
+        let result = check_parse_template(template_str);
+        assert!(matches!(
+            result,
+            CheckParseTemplateResult::Error { errors: _ }
         ));
     }
 }